Schemas for MongodB

+++ Database/Mongo.js
@@ -2,6 +2,10 @@ const { MongoClient, ServerApiVersion, ObjectId,  } = require('mongodb');
 
 const uri = `mongodb+srv://${process.env.DATABASE_USER}:${process.env.DATABASE_PASW}@${process.env.DATABASE_NAME}.8t4t4k6.mongodb.net/?retryWrites=true&w=majority`;
 
+const { CustomerMongoSchema } = require('./Models/Customer.schema');
+const { PetMongoSchema } = require('./Models/Pet.schema');
+const { MedicineMongoSchema } = require('./Models/Medicine.schema');
+
 class MongoDB {
   
   // Generar el cliente de MongoDB para aplicar el patrón Singleton
@@ -34,13 +38,45 @@ class MongoDB {
 
   /** METODOS */
 
-  /** BACKUP - EXCEL */
-  async dataExcel(data, collection) {
-
+  async CreateCollections () {
+    
     return this.connect().then((db) => {
 
-      db.dropCollection(collection);
+      
+
+      
+      db.createCollection("Mascotas", PetMongoSchema);
+      db.createCollection("Medicamentos", MedicineMongoSchema);
+      
+    });
+  }
+
+  /** BACKUP - EXCEL */
+  async dataExcel(data, collection) {
+    return this.connect().then(async (db) => {
+
+      let Collections = await db.listCollections().toArray();
+      Collections = Collections.map((item) => {
+        return item.name;
+      })
+
+      if (Collections.includes(collection)) {
+        await db.dropCollection(collection);
+      }
+
+      switch (collection) {
+        case "Clientes":
+          await db.createCollection(collection, CustomerMongoSchema);
+          break;
+        case "Mascotas":
+          await db.createCollection(collection, PetMongoSchema);
+          break;
+        case "Medicamentos":
+          await db.createCollection(collection, MedicineMongoSchema);
+          break;
+      }
       return db.collection(collection).insertMany([...data]);
+
     });
 
   }
